gas-sensor/requirements.txt
termcolor
requests
dropbox
RPi.GPIO
